Support and test server restart, dynamic membership

+++ cloudata-server-shared/src/main/java/com/cloudata/TestUtils.java
@@ -6,6 +6,10 @@ import com.google.protobuf.ByteString;
 
 public class TestUtils {
     public static void rmdir(File dir) {
+        if (!dir.exists()) {
+            return;
+        }
+
         for (File f : dir.listFiles()) {
             if (f.isFile()) {
                 f.delete();
